+++ src/color.rs
@@ -4,10 +4,16 @@ use palette::{rgb::Rgb, FromColor, Hsl, IntoColor, Srgb};
 pub(crate) struct Color {
     pub(crate) associated_pure_color: PureColor,
     pub(crate) value: Srgb<u8>,
-    pub(crate) distance: u32,
+    pub(crate) distance: f64,
 }
 
 impl Color {
+    /// Create a new color
+    /// The distance is calculated using the Euclidean distance formula
+    ///
+    /// # Arguments
+    /// * `pure_color` - A PureColor enum
+    /// * `value` - A Srgb<u8> color
     pub(crate) fn new(pure_color: PureColor, value: Srgb<u8>) -> Self {
         let distance = Color::get_distance(&Color::from(pure_color).value, &value);
 
@@ -18,14 +24,16 @@ impl Color {
         }
     }
 
+    /// Create a new color from a pure color
     pub(crate) fn from(pure_color: PureColor) -> Self {
         Color {
             associated_pure_color: pure_color,
             value: pure_color.get_rgb(),
-            distance: 0,
+            distance: 0.0,
         }
     }
 
+    /// Get the inverse of the color
     pub(crate) fn get_inverse(&self) -> Self {
         let rgb_color_inverse = Srgb::new(
             255 - self.value.red,
@@ -37,29 +45,36 @@ impl Color {
         Color::new(pure_color_inverse, rgb_color_inverse)
     }
 
-    // Order of c1 and c2 doesn't matter
-    pub(crate) fn get_distance(c1: &Srgb<u8>, c2: &Srgb<u8>) -> u32 {
+    /// Get the distance between two colors
+    /// The distance is calculated using the Euclidean distance formula
+    ///
+    /// # Arguments
+    /// * `c1` - A reference to a Srgb<u8> color
+    /// * `c2` - A reference to a Srgb<u8> color
+    pub(crate) fn get_distance(c1: &Srgb<u8>, c2: &Srgb<u8>) -> f64 {
+        // Order of c1 and c2 doesn't matter
         let dr = c1.red as i32 - c2.red as i32;
         let dg = c1.green as i32 - c2.green as i32;
         let db = c1.blue as i32 - c2.blue as i32;
 
-        (dr * dr + dg * dg + db * db) as u32
+        ((dr * dr + dg * dg + db * db) as f64).sqrt()
     }
 
+    /// Convert the color to a hex string
     pub(crate) fn to_hex(self) -> String {
         let (r, g, b) = self.value.into_components();
 
         format!("{:02X}{:02X}{:02X}", r, g, b)
     }
 
+    /// Saturate the color
+    /// The percentage is squared to make the saturation effect more noticeable
+    ///
+    /// # Arguments
+    /// * `percentage` - A f32 value between 0.0 and 1.0
     pub(crate) fn to_saturated(mut self, percentage: f32) -> Self {
-        let percentage = if percentage > 1.0 { 1.0 } else { percentage };
-        let rgb: Rgb = Rgb::new(
-            self.value.red as f32 / 255.0,
-            self.value.green as f32 / 255.0,
-            self.value.blue as f32 / 255.0,
-        );
-        let hsl: Hsl = Hsl::from_color(rgb);
+        let percentage = percentage.clamp(0.0, 1.0);
+        let hsl: Hsl = Hsl::from_color(self.value.into_format::<f32>());
         let updated_saturation: Hsl = Hsl::new(
             hsl.hue,
             hsl.saturation * percentage * percentage,
@@ -75,9 +90,30 @@ impl Color {
 
         self
     }
+
+    /// Add lightness to the color
+    ///
+    /// # Arguments
+    ///
+    /// * `value` - A f32 value between 0.0 and 1.0
+    ///
+    pub(crate) fn add_lightness(mut self, value: f32) -> Self {
+        let hsl: Hsl = Hsl::from_color(self.value.into_format::<f32>());
+        let updated_lightness = (hsl.lightness + value.clamp(0.0, 1.0)).clamp(0.0, 1.0);
+        let hsl: Hsl = Hsl::new(hsl.hue, hsl.saturation, updated_lightness);
+        let updated_rgb: Rgb = hsl.into_color();
+
+        self.value = Srgb::new(
+            (updated_rgb.red * 255.0) as u8,
+            (updated_rgb.green * 255.0) as u8,
+            (updated_rgb.blue * 255.0) as u8,
+        );
+
+        self
+    }
 }
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub(crate) enum PureColor {
     Red,
     Yellow,
@@ -143,5 +179,42 @@ impl PureColor {
             PureColor::SpringGreen => PureColor::Purple,
             PureColor::LightCyan => PureColor::Brown,
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_add_lightness() {
+        let color = Color::new(PureColor::Red, Srgb::new(255, 0, 0));
+        let color = color.add_lightness(0.1);
+
+        assert_eq!(color.value, Srgb::new(255, 51, 51));
+    }
+
+    #[test]
+    fn test_get_distance() {
+        let color1 = Srgb::new(255, 0, 0);
+        let color2 = Srgb::new(0, 255, 0);
+
+        assert_eq!(Color::get_distance(&color1, &color2), 360.62445840513925);
+    }
+
+    #[test]
+    fn test_get_inverse() {
+        let color = Color::new(PureColor::Red, Srgb::new(255, 0, 0));
+        let color = color.get_inverse();
+
+        assert_eq!(color.associated_pure_color, PureColor::Cyan);
+        assert_eq!(color.value, Srgb::new(0, 255, 255));
+    }
+
+    #[test]
+    fn test_to_hex() {
+        let color = Color::new(PureColor::Red, Srgb::new(255, 0, 0));
+
+        assert_eq!(color.to_hex(), "FF0000");
     }
 }